+++ mobile/android/android-components/components/compose/cfr/src/main/java/mozilla/components/compose/cfr/helper/DisplayOrientationListener.kt
@@ -7,6 +7,7 @@ package mozilla.components.compose.cfr.helper
 import android.content.Context
 import android.hardware.display.DisplayManager
 import android.hardware.display.DisplayManager.DisplayListener
+import android.os.Build
 import androidx.annotation.VisibleForTesting
 
 /**
@@ -21,13 +22,13 @@ import androidx.annotation.VisibleForTesting
  * No updates will be triggered if the "Auto-rotate" functionality is disabled for the device.
  */
 internal class DisplayOrientationListener(
-    context: Context,
+    private val context: Context,
     val onDisplayRotationChanged: () -> Unit,
 ) : DisplayListener {
     private val displayManager = context.getSystemService(Context.DISPLAY_SERVICE) as DisplayManager
 
     @VisibleForTesting
-    internal var currentOrientation = displayManager.displays[0].rotation
+    internal var currentOrientation = getCurrentOrientation()
 
     /**
      * Start listening for display orientation changes.
@@ -49,11 +50,16 @@ internal class DisplayOrientationListener(
     override fun onDisplayRemoved(displayId: Int) = Unit
 
     override fun onDisplayChanged(displayId: Int) {
-        val display = displayManager.getDisplay(displayId) ?: return // getDisplay may return null
+        val newOrientation = getCurrentOrientation(displayId)
 
-        if (display.rotation != currentOrientation) {
-            currentOrientation = display.rotation
+        if (newOrientation != this.currentOrientation) {
+            this.currentOrientation = newOrientation
             onDisplayRotationChanged()
         }
+    }
+
+    private fun getCurrentOrientation(displayId: Int = 0): Int = when (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+        true -> context.resources.configuration.orientation
+        false -> displayManager.getDisplay(displayId)?.rotation ?: currentOrientation
     }
 }